// nextjs/src/lib/types/audit-log.ts
export enum AuditAction {
    CREATE = 'CREATE',
    UPDATE = 'UPDATE',
    DELETE = 'DELETE',
    READ = 'READ',
    LOGIN = 'LOGIN',
    LOGOUT = 'LOGOUT',
}

export enum AuditEntity {
    USER = 'USER',
    WORKSPACE = 'WORKSPACE',
    BOARD = 'BOARD',
    LIST = 'LIST',
    CARD = 'CARD',
    MEMBER = 'MEMBER',
}